refactors channel and message types to view and model folders
