+++ ui-src/src/cells/holo-chat/components/messages/messages.story.tsx
@@ -10,7 +10,7 @@ import listMessages from './listMessages.md'
 import reply from './reply.md'
 import CreateStore from '../../../../store'
 import  * as constants from '../../constants'
-import{Message as MessageType} from '../../types/message'
+import{Message as MessageType} from '../../types/view/message'
 
 
 configure({adapter: new Adapter()})
