+++ ui-src/src/cells/holo-chat/reducer.ts
@@ -1,32 +1,39 @@
 import { combineReducers } from 'redux'
 import { ActionType, getType } from 'typesafe-actions';
 import * as chatActions from './actions'
-import {Channel} from './types/channel'
+import {Channel} from './types/model/channel'
+import {Message} from './types/model/message'
+
 
 // create a union type that is all possible chat actions
 export type ChatAction = ActionType<typeof chatActions>;
 
 export interface HoloChatState {
-  myChannels: Array<Channel>
+  myChannels: Array<Channel>,
+  currentMessages: Array<Message>
 }
 
 export const initialState: HoloChatState = {
-  myChannels: []
+  myChannels: [],
+  currentMessages: []
 }
 
 
 export function holochatReducer (state = initialState, action: ChatAction) {
   switch (action.type) {
     case getType(chatActions.CreateCustomChannel.success):
-      	console.log("create custom channel action triggered!");
       	return state // do nothing for now
     case getType(chatActions.AddMembers.success):
-    	console.log("add members action triggered")
-    	return state
+    	return state // do nothing for now
     case getType(chatActions.GetMyChannels.success):
     	return {
         ...state,
         myChannels: action.payload.data
+      }
+    case getType(chatActions.GetMessages.success):
+      return {
+        ...state,
+        currentMessages: action.payload.data
       }
     default:
       return state
