+++ ui-src/src/cells/holo-chat/types/model/channel.ts
@@ -1,0 +1,8 @@
+
+
+export interface Channel {
+	id: string,
+	name: string,
+	description: string
+}
+
