+++ ui-src/src/cells/holo-chat/types/model/message.ts
@@ -1,0 +1,10 @@
+
+export interface Message {  
+	author: string,
+	timestamp: string,
+	content: {
+		text: string,
+		mediaLink?: string
+	},
+	channelId: string 
+}
