+++ ui-src/src/cells/holo-chat/types/view/channel.ts
@@ -1,0 +1,1 @@
+export type Channel = any
