+++ ui-src/src/cells/holo-chat/types/view/message.ts
@@ -1,0 +1,5 @@
+export type Message = any
+
+export enum MessageType {
+	CHAT
+}
